Move application_session.h so it can be included by the frontend
