+++ src/include/server/mir/scene/output_properties_cache.h
