+++ src/server/scene/CMakeLists.txt
@@ -31,7 +31,6 @@ ADD_LIBRARY(
   unsupported_coordinate_translator.cpp
   timeout_application_not_responding_detector.cpp
   output_properties_cache.cpp
-  output_properties_cache.h
   application_not_responding_detector_wrapper.cpp
   ${CMAKE_SOURCE_DIR}/include/server/mir/scene/surface_observer.h
 )
